+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - fix: Access Token introspection was not working with a public client. Fixed.
 - feat: Add support for JWT Access Token (helm values: `oidc.jwtAccessToken: true|false`  )
 - feat: Support of the Client credential flow
-- feat: clientSecret can now also be stored in a k8s secret
+- feat: HashedSecret has been removed. A list of secrets can be provided instead, thus allowing secrets rotation
+- feat: kc hash subcommand has been modified